LayoutTests/imported/w3c/web-platform-tests/html/user-activation/message-event-init.tentative-expected.txt


PASS MessageEventInit user activation set
FAIL MessageEventInit user activation not set assert_equals: userActivation attribute expected (object) null but got (undefined) undefined
